pyproject.toml
[project]
name = "vms-backend"
version = "1.0.0"
description = "Video Management System Backend with AI Integration"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "pydantic==2.5.0",
    "websockets==12.0",
    "python-multipart==0.0.6",
    "anthropic>=0.28.0",
    "python-dotenv==1.0.0",
    "pillow>=10.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.25.0"
]
